Update TableEntry format: TableEntries are now seperate from IR
