+++ p4_symbolic/ir/BUILD.bazel
@@ -60,6 +60,8 @@ cc_library(
         ":ir_cc_proto",
         "//p4_symbolic/util",
         "@com_google_absl//absl/strings",
+        "@p4_pdpi//p4_pdpi:pdpi",
+        "@p4_pdpi//p4_pdpi/utils:status_utils",
     ],
 )
 
@@ -79,6 +81,7 @@ cc_library(
         "@com_google_absl//absl/strings",
         "@com_google_protobuf//:protobuf_headers",
         "@p4_pdpi//p4_pdpi:pdpi",
+        "@p4_pdpi//p4_pdpi/utils:status_utils",
     ],
 )
