+++ p4_symbolic/ir/expected/table.pb.txt
@@ -290,16 +290,6 @@ tables {
       size: 1024
     }
     table_implementation {
-      entries {
-        action: "MyIngress.set_egress_spec"
-        match_values: 0
-        action_parameters: 1
-      }
-      entries {
-        action: "MyIngress.set_egress_spec"
-        match_values: 1
-        action_parameters: 0
-      }
     }
   }
 }
